+++ temba/flows/models.py
@@ -24,6 +24,7 @@ from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
 
 from temba import mailroom
+from temba.classifiers.models import Classifier
 from temba.assets.models import register_asset_store
 from temba.channels.models import Channel, ChannelConnection
 from temba.contacts.models import URN, Contact, ContactField, ContactGroup
@@ -302,6 +303,8 @@ class Flow(TembaModel):
     channel_dependencies = models.ManyToManyField(Channel, related_name="dependent_flows")
 
     label_dependencies = models.ManyToManyField(Label, related_name="dependent_flows")
+
+    classifier_dependencies = models.ManyToManyField(Classifier, related_name="dependent_flows")
 
     @classmethod
     def create(
@@ -1018,6 +1021,7 @@ class Flow(TembaModel):
         dependencies.update(self.flow_dependencies.all())
         dependencies.update(self.field_dependencies.all())
         dependencies.update(self.group_dependencies.all())
+        dependencies.update(self.classifier_dependencies.all())
         return dependencies
 
     def is_legacy(self):
@@ -1679,6 +1683,7 @@ class Flow(TembaModel):
         group_uuids = [g["uuid"] for g in dependencies.get("groups", [])]
         channel_uuids = [g["uuid"] for g in dependencies.get("channels", [])]
         label_uuids = [g["uuid"] for g in dependencies.get("labels", [])]
+        classifier_uuids = [c["uuid"] for c in dependencies.get("classifiers", [])]
 
         # still need to do lazy creation of fields in the case of a flow import
         if len(field_keys):
@@ -1701,6 +1706,7 @@ class Flow(TembaModel):
         groups = ContactGroup.user_groups.filter(org=self.org, uuid__in=group_uuids)
         channels = Channel.objects.filter(org=self.org, uuid__in=channel_uuids, is_active=True)
         labels = Label.label_objects.filter(org=self.org, uuid__in=label_uuids)
+        classifiers = Classifier.objects.filter(org=self.org, uuid__in=classifier_uuids)
 
         self.field_dependencies.clear()
         self.field_dependencies.add(*fields)
@@ -1716,6 +1722,9 @@ class Flow(TembaModel):
 
         self.label_dependencies.clear()
         self.label_dependencies.add(*labels)
+
+        self.classifier_dependencies.clear()
+        self.classifier_dependencies.add(*classifiers)
 
     def release(self):
         """